feat: 完成所有基本接口 - 完成用户的会议室推荐 - 完成用户订单创建、订单支付、取消申请、取消取消申请 - 完成员工、管理员的取消申请审核 - 完成员工的会议室状态切换 - 完成用户的余额充值 - 完成系统鉴权报错返回 - 完善文档与数据库建表语句

+++ backend/src/main/java/cn/polister/dianmeetingsystem/entity/RoomOrder.java
@@ -1,5 +1,6 @@
 package cn.polister.dianmeetingsystem.entity;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import java.io.Serializable;
@@ -32,7 +33,7 @@ public class RoomOrder  {
 
     private Date endTime;
 
-    private Double totalPrice;
+    private BigDecimal totalPrice;
 
     private String orderStatus;
 
@@ -44,7 +45,7 @@ public class RoomOrder  {
 
     private Date cancelTime;
 
-    private Double refundAmount;
+    private BigDecimal refundAmount;
 
     private Integer delFlag;
